cosmetic changes
  1. Single possibility to make fiber dead is resume it.
Considering EM.cancel_timeout do not raise error, we
absolutely sure: if fiber is not alive, than it is not
among sleepers.
  2. lets Mutex#sleep accept block, which will be executed
just after Fiber resumes, but before it will try to lock a
mutex. It helps us to not signal to thread, already waken
up a timer.
  Another possibility to prevent it is to check mutex's
sleepers. I could not tell which way is better, I chose
this one.